Remove `include/grpcpp/impl/codegen/proto_buffer_writer.h` (#31268)
